rd2 New Camera Top
Until now we were developing an unique top part for the robot. After
having some problems due to the diferent configurations that is posible
to have, we change the concept. Now, the top of the robot we will made
of several parts that can be combined in order to support more
hardware(raspberry pi,odrid,beaglebone).